Fixed return type.
git-svn-id: file:///home/svn/meshLib@211 a422a631-36e5-491f-99c6-f798677b70e4

+++ include/meshLib/trn.hpp
@@ -152,12 +152,12 @@ namespace ml
     unsigned int readTRN( std::istream &file,
 			  const std::string &debugString = "" );
 
-    const float &getHeight() const
+    const unsigned int &getHeight() const
     {
       return height;
     }
 
-    const float &getWidth() const
+    const unsigned int &getWidth() const
     {
       return width;
     }
